Fetch data and render it at the rockets page
