--- src/pages/rockets-page/rockets-page.component.jsx
@@ -1,13 +1,0 @@
-import React from "react";
-import RocketsList from "../../components/rockets/rocketsList/RocketsList.component";
-
-const RocketsPage = () => {
-  return (
-  <div>
-    <h1>Rockets</h1>
-    <RocketsList />
-  </div>
-  );
-}
-
-export default RocketsPage;
