+++ src/pages/rockets-page/rockets-page.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import RocketsList from '../../components/rockets/rocketsList/RocketsList';
+
+const RocketsPage = () => (
+  <div>
+    <h1>Rockets</h1>
+    <RocketsList />
+  </div>
+);
+
+export default RocketsPage;
